% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project Tools# get_or_set_dir.R
\name{function_catcher}
\alias{function_catcher}
\title{Function Catcher}
\usage{
function_catcher(eachfile, final_function_table)
}
\arguments{
\item{eachfile}{path of the script file}

\item{final_function_table}{the final table of functions.}
}
\description{
Function Catcher
}
\details{
catch the functions in the scripts.
sub function for the load_functions
}
