% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data Process Tools# get_colors.R
\name{get_colors}
\alias{get_colors}
\title{get colors and breaks for plot(usually for continue var in heatmap)}
\usage{
get_colors(
  max = 3,
  min = -3,
  breaks = 0.05,
  print_notes = F,
  white_color_expend = 2,
  blue_raw = c("#5000B8", "#2A52BE", "#1B53DE", "#72AEEE"),
  red_raw = c("#FFFAB3", "#FFBF00", "#EB4E00", "#DF0029", "#C50023", "#8B0016"),
  zero_for_balance = 0,
  balance_color = "white"
)
}
\arguments{
\item{max}{the top end of the data to be mapped.}

\item{min}{the below end of the data to be mapped.}

\item{breaks}{the breaks for diff colors.}

\item{print_notes}{wether to print the demo codes for users}

\item{white_color_expend}{the length expand round the zero point.}

\item{blue_raw}{the blue colors}

\item{red_raw}{the red colors}

\item{zero_for_balance}{the colors between the 'balance color' and other color.}

\item{balance_color}{the color in the middle.}
}
\value{
a list with two sub vector

result$my_color the colors vector.

result$my_breaks the breaks vector.
}
\description{
get colors and breaks for plot(usually for continue var in heatmap)
}
\examples{
\dontrun{
result <- get_colors(max = 3,min = -3,
    breaks = 0.05,
    white_color_expend = 2)
}
}
