% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap in gg# 03 get_tree_plot.R
\name{get_heatmap_tree}
\alias{get_heatmap_tree}
\title{get ggtree form ExpressionSet}
\usage{
get_heatmap_tree(
  ExpressionSet_input,
  heatmap_ls,
  tree.size = 3,
  tree.color = "red",
  selected_var.feature = "type",
  selected_var.pheno = "Species",
  heighted_v = c("Petal.Length", "Petal.Width"),
  tip.shape = 21,
  tip.size = 8,
  colors_c_readonly = colors(distinct = T) \%>\% .[c(6, 325, 3, 400, 5, 7, 29, 38, 356,
    129, 76, 293, 156)]
)
}
\arguments{
\item{ExpressionSet_input}{输入数据集}

\item{heatmap_ls}{a list to store the heatmap subplots.}

\item{tree.color}{color for the tree.}

\item{heighted_v}{the nodes need to be heighted in branch.}

\item{tip.shape}{the shape of the nodes}

\item{tip.size}{the size of the nodes}

\item{colors_c_readonly}{the colors for mapping.#'}

\item{heighted_nodes}{要进行重点强调的Nodes ## 一般不建议与叶子的图形强调一同使用。}

\item{as_tip_fill.feature}{character ## 要作为填充依据的表型列名}

\item{as_tip_fill.pheno}{要作为填充色依据的特征列名。}
}
\value{
heatmap_ls 返回包含左、上两棵树的图片列表结果。
}
\description{
像 ggtree 创建的Tree对象添加数据的方法：创建一个数据框。
确保第一列是树的根节点的名字，其他的列为各种属性即可。
}
\details{
使用ggtree绘制树形图。
需要实现的功能：
绘制出树形结构带/不带tip.label
调节横向/纵向的间距
同时产出两个树形图（横、竖方向）
调整树的颜色、粗细、线性
调整树的叶子的颜色、形状。
}
\examples{
\dontrun{
library(pacman)
p_load(Loafer)
p_load(ggtree)
p_load(ggplot2)

get_or_set_dir()

heatmap_ls <- get_heatmap_tree(ExpressionSet_input = ExpressionSet,
                 heatmap_ls = heatmap_ls,
                 tree.size = 3,
                 tree.color = "red",
                 selected_var.feature = NULL, ##  鐢ㄦ潵寮鸿皟鐨勫彾鑺傜偣
                 selected_var.pheno = NULL,
                 heighted_v = NULL,
                 tip.shape = 21,
                 tip.size = 8,
                 colors_c_readonly = colors(distinct = T) \%>\% `[`
                 (.,c(6,325,3,400,5,7,29,38,356, 129, 76, 293, 156)))
heatmap_ls$Left_Tree
heatmap_ls$Top_Tree
}
}
