% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_01_Get_module_data.R
\name{get_ls_for_model}
\alias{get_ls_for_model}
\title{Turn the ExperimentSet S4 into a data frame}
\usage{
get_ls_for_model(set)
}
\arguments{
\item{set}{train_set or test_set, ExperimentSet. process and get the new dat for module.}
}
\value{
a dataframe with "group" and other var.
}
\description{
Turn the ExperimentSet S4 into a data frame
}
