% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation_01_ Get_ls_from_exl_5.0.R
\name{Get_ls_from_exl_5.0}
\alias{Get_ls_from_exl_5.0}
\title{Get ls from exl module.
从指定格式的模板文件中获取注释。
进行相关分析的函数汇总。
主要针对菌群数据做了一定的优化，化简需要的操作，并且自动处理一部分问题。}
\usage{
Get_ls_from_exl_5.0(
  file_name = "filename for module.xlsx",
  sep_code = "|",
  adj_index = 1,
  remove_letter__ = T,
  extract_method = "sig in group",
  match_mode = "cleaned_match",
  get_level_by = "sep_code",
  IS_microbiota = F
)
}
\arguments{
\item{real_I}{the real group of a each sample in dataset}

\item{pred_num}{the predicted num of each sample}

\item{fit_note}{the note to be used in the filename to seperate the results}

\item{IS_smooth}{wether to smooth the ROC lines.}
}
\description{
Get ls from exl module.
从指定格式的模板文件中获取注释。
进行相关分析的函数汇总。
主要针对菌群数据做了一定的优化，化简需要的操作，并且自动处理一部分问题。
}
\examples{
\dontrun{
pred_num <- c(0.9,0.1,0.33,0.87)
real_I <- c(0,0,1,1)
plot_roc(real_I = real_I, pred_num = pred_num)
}
}
