% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_or_set_dir.R
\name{get_or_set_dir}
\alias{get_or_set_dir}
\title{get or set the dir by Rstudio
Get the dir where the script were lying in.
Set the dir as main_path and create some sub file folder inside the dir.
Chose one file as the new working dir.}
\usage{
get_or_set_dir(
  main_trunk_as_wd = F,
  source_path = "0_data_source/",
  output_path = "2_output/",
  other_dirs = c(function_path = "DIY Function/"),
  setwd_in = source_path
)
}
\arguments{
\item{main_trunk_as_wd, }{Boolean, IF T, the main dir will be set as new wd.}

\item{source_path, }{character, where raw data will be stored in a project.}

\item{output_path, }{character, where the output of an analysis will be saved.}

\item{other_dirs, }{character vector with var name. the name will be assigned and be returned to global env.}

\item{setwd_in, }{name, source path default. when the main_trunk_as_wd is set as F, this var will be set as new wd.}
}
\value{
return a summary about working dir and the source and output dir.
}
\description{
get or set the dir by Rstudio
Get the dir where the script were lying in.
Set the dir as main_path and create some sub file folder inside the dir.
Chose one file as the new working dir.
}
\examples{
\dontrun{# set the dir as wd:
get_or_set_dir(main_trunk_as_wd = T,
   source_path = "0_data_source/",
   output_path = "2_output/",
   other_dirs = c(function_path = "DIY Function/", other_dir_test = "Test dir"),
   setwd_in = source_path)
# set the source as wd:
get_or_set_dir(main_trunk_as_wd = F,
   source_path = "0_data_source/",
   output_path = "2_output/",
   other_dirs = c(function_path = "DIY Function/", Dir_test = "Test dir"),
   setwd_in = source_path)
}
}
