% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_01_Get_module_data.R
\name{get_seeds}
\alias{get_seeds}
\title{get random seeds for random selection}
\usage{
get_seeds(Today = Sys.Date(), length = 100)
}
\arguments{
\item{Today}{the date for seed}

\item{length}{the num of seeds needed}
}
\value{
seeds
}
\description{
get random seeds for random selection
}
\examples{
\dontrun{
seeds <- get_seeds(Today = Sys.Date() + 3, length = 10)
}
}
