% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_02_Build_modules_and_get_AUC.R
\name{get_the_auc.Lasso_and_Ridge}
\alias{get_the_auc.Lasso_and_Ridge}
\title{building modules by using LR 使用LR进行建模
为Lasso 和 Ridge 模型提供AUC}
\usage{
get_the_auc.Lasso_and_Ridge(module_obj)
}
\arguments{
\item{module_obj}{inheritParams from get_the_auc}
}
\description{
building modules by using LR 使用LR进行建模
为Lasso 和 Ridge 模型提供AUC
}
