% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_02_Build_modules_and_get_AUC.R
\name{get_the_auc}
\alias{get_the_auc}
\title{get the auc of train and test data.}
\usage{
get_the_auc(module_obj, method = "normal")
}
\arguments{
\item{module_obj}{the obj for the moduling.}

\item{method}{to call different predict for different method; for some fit need different predict param. such as Bayes et.al.}
}
\value{
res_df a data.frame contains method_note: a note about the module and method and param used.
}
\description{
get the auc of train and test data.
可以为RF、LR、贝叶斯、弹性网络等模型提供AUC。
}
