% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data Process Tools# get Z Score.R
\name{get_z_score}
\alias{get_z_score}
\title{Get Z score from matrix or data.frame(pure number.)}
\usage{
get_z_score(m = df, by = "col")
}
\arguments{
\item{m}{a mun matrix or data.frame to be scaled.}

\item{by}{specified "col" or "row" to be scaled.}
}
\value{
a matrix/data.frame with scaled value.
}
\description{
Get Z score from matrix or data.frame(pure number.)
}
\details{
使用纯数据数据框或者矩阵做为输入，选择按照行或列进行Z得分转换
}
\examples{
\dontrun{
get_z_score(m = iris[,1:4], by = row)
get_z_score(m = iris[,1:4], by = "row")
get_z_score(m = iris[,1:4], by = "col")
get_z_score(m = iris[,1:4] \%>\% as.matrix(), by = "row")
get_z_score(m = iris[,1:4] \%>\% as.matrix(), by = "col")
}
}
