% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project Tools# those not functions.R
\name{\%notin\%}
\alias{\%notin\%}
\title{not.in}
\usage{
... \%notin\% NA
}
\description{
not.in
}
\details{
find out items not contained in sth.
reverse side of the \%in\%
\subsection{中缀运算符。}{

在R包中创建中缀运算函数
}
}
\examples{
\dontrun{
"A" \%notin\% "B"
"A" \%notin\% "A"
"A" \%notin\% LETTERS
}
}
