% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Greek_letters_to_Enl.R
\name{Greek_to_Enl}
\alias{Greek_to_Enl}
\title{Trans Greek letters to English.}
\usage{
Greek_to_Enl(input, Greek_to_Enl_df = Greek_to_Enl_df)
}
\arguments{
\item{input, }{vector needs to trans into English}

\item{Greek_to_Enl_df, }{dictionary for trans.}
}
\value{

}
\description{
Trans Greek letters to English.
}
\examples{
\dontrun{
 Greek_to_Enl(input = c("螕", "螖", "未"), Greek_to_Enl_df = Greek_to_Enl_df)
}
}
