% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_MP_datas.R
\name{import_MP_datas}
\alias{import_MP_datas}
\title{Get S4 obj ExperimentData from MP data file directly}
\usage{
import_MP_datas(
  Allmet_file,
  group_info_file,
  obj_note,
  remove_QC = c("QC", "qc", "Qc"),
  ClassNote_involved_c = phenoData$ClassNote \%>\% unique()
)
}
\arguments{
\item{Allmet_file}{allmet data file name}

\item{group_info_file}{group info file name}

\item{obj_note}{note for the obj}

\item{remove_QC}{charactor vector, remove the QC if necessary.}

\item{ClassNote_involved_c}{charactor vector, select groups need to be imports}

\item{groups_involved_c}{groups involved in the analysis.}
}
\value{
返回一个Experiment 的S4对象
}
\description{
Get S4 obj ExperimentData from MP data file directly
}
\details{
从麦特绘谱的结果文件直接生成S4对象。

depracted param Folder_path 若给出文件夹所在的位置。 直接从中解析出需要的数据,感觉用处不大，不够灵活
param Treat_involved  涉及到的Treats，不够灵活，手动搞定文件吧
}
\examples{
\dontrun{
## 鎼炲嚑涓暟鎹枃浠躲€傛斁鍦ㄥ伐浣滅洰褰曚腑锛岀劧鍚庢墽琛屼笅杩颁唬鐮侊細
getwd()
import_MP_datas(Allmet_file = "04_AllMet 璁粌闆<86>.csv",
    group_info_file = "Class_Info 璁粌闆<86>.csv",
    obj_note = "The note")
exprs_m <- ExpressionSet@assayData$exprs ## expr data.
pheno_df <- ExpressionSet@phenoData@data ## phenoData data.
feature_df <- ExpressionSet@featureData@data ## feature data.

}
}
