% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project Tools# get_or_set_dir.R
\name{load_functions}
\alias{load_functions}
\title{scripts_sourcer source scripts and load local functions into Global.env.}
\usage{
load_functions(
  my_function_path = fun_path,
  functions_collection_path = main_path,
  safe_word = "# be quiet",
  repl_safe_word = "#\\\\s*silence",
  encoding = "utf-8"
)
}
\arguments{
\item{my_function_path}{the path in which your functions and env initialization scripts stored.}

\item{functions_collection_path}{the path for the functions collection xlsx file.}

\item{safe_word}{character, scripts with the 'safe_word' won't be sourced.}

\item{repl_safe_word}{character, safe_word pass to grepl function like: grepl(repl_safe_word, filename)}

\item{encoding}{encoding of the scripts}
}
\value{
final_function_table, a table of Functions.
}
\description{
scripts_sourcer source scripts and load local functions into Global.env.
}
\details{
指定（绝对路径或相对当前目录的相对路径）某个文件夹
该文件夹及其子文件夹内的脚本将按照字母顺序全部依次加载运行。
各个函数会被加载到环境变量中。脚本中的初始化设置也会起效
若想不加载某些内容。可以设定安全词，这样具有安全词的脚本边不会被运行。
}
\examples{
\dontrun{
get_or_set_dir(other_dirs = c(fun_path = "DIY Function"))
load_functions(my_function_path = fun_path, functions_collection_path = main_path)
}
}
