% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation_03_Meji_to_human
%   美吉16S菌群数据转换成人看的.R
\name{meiji_otu_to_human}
\alias{meiji_otu_to_human}
\title{trans meiji(美吉) 16S OTU data into normal data}
\usage{
meiji_otu_to_human(xlsx.filename = "文件名.xlsx")
}
\arguments{
\item{xlsx.filename}{file name of the xlsx data.}
}
\description{
trans meiji(美吉) 16S OTU data into normal data
}
\details{
Substract
主要是将几种无明确分类的部分去掉，并且将各个层级的名称进行合并
同时还将重复的物质进行了去重
此外，可能会有一个本来就是空的（从第二层开始就鉴定不到东西），OTU，会被自动去掉。
}
