% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_02_Build_modules_and_get_AUC.R
\name{module.Bayes}
\alias{module.Bayes}
\title{Module with Bayes采用贝叶斯分类器进行计算}
\usage{
module.Bayes(module_obj, laplace = 0)
}
\arguments{
\item{module_obj}{inheritParams from get_the_auc}

\item{laplace}{param for smooth.}
}
\description{
building modules by using LR 使用LR进行建模
}
