% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_02_Build_modules_and_get_AUC.R
\name{module.GB}
\alias{module.GB}
\title{building models by using gbm from gbm package}
\usage{
module.GB(module_obj)
}
\arguments{
\item{module_obj}{inheritParams from get_the_auc}
}
\description{
\subsection{当预测分组全部相同时会有问题导致用不了}{
}
}
