% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_02_Build_modules_and_get_AUC.R
\name{module.Lasso_or_Ridge}
\alias{module.Lasso_or_Ridge}
\title{get module by Lasso_or_Ridge使用岭回归进行分类分析。}
\usage{
module.Lasso_or_Ridge(module_obj, Lasso_or_Ridge = "Lasso")
}
\arguments{
\item{module_obj}{inheritParams from get_the_auc}

\item{Lasso_or_Ridge}{character: Lasso or Ridge to set different methods.}
}
\description{
参考资料： https://www.pluralsight.com/guides/linear-lasso-and-ridge-regression-with-r
}
