% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_02_Build_modules_and_get_AUC.R
\name{module.LR}
\alias{module.LR}
\title{building modules by using LR 使用LR进行建模
本页的函数是为了简化实际筛选模型/方法/参数以便获得更好的预测效果所做的尝试
具体修改的主要有：同时提供训练集和验证集的AUC结果；部分方法会提供必要的参数；尝试了多种不同的方法。}
\usage{
module.LR(module_obj)
}
\description{
building modules by using LR 使用LR进行建模
本页的函数是为了简化实际筛选模型/方法/参数以便获得更好的预测效果所做的尝试
具体修改的主要有：同时提供训练集和验证集的AUC结果；部分方法会提供必要的参数；尝试了多种不同的方法。
}
