% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neat_round.R
\name{neat_round}
\alias{neat_round}
\title{works like format function.}
\usage{
neat_round(num, digits = 3)
}
\arguments{
\item{num}{numeric, vector, num need to be round.}

\item{digits, }{digits of the data rounded.}
}
\description{
works like format function.
}
\details{
round the data and keep the data with selected length.
zero in the tail will also be keeped.
}
\examples{
\dontrun{
num <- c(1,3,5.2,-2.3)
neat_round(num, digits=3)#'
}
}
