% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project Tools# those not functions.R
\name{not.grepl}
\alias{not.grepl}
\title{not.grepl}
\usage{
not.grepl(...)
}
\arguments{
\item{x}{quary vector}
}
\description{
not.grepl
}
\details{
find out items not contains sth.
reverse side of the grepl
}
\examples{
\dontrun{
"The belowing two equals while the first can save one brackets, cause the trd line(without bracket) will cause wrong."
not.grepl("F",LETTERS) \%>\% LETTERS[.]
(grepl("F",LETTERS) == F) \%>\% LETTERS[.]
try(grepl("F",LETTERS) == F \%>\% LETTERS[.])
}
}
