% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project Tools# not functions.R
\name{not.na}
\alias{not.na}
\title{not.na}
\usage{
not.na(x)
}
\arguments{
\item{x}{quary vector}
}
\description{
not.na
}
\details{
find out items not NA
}
\examples{
\dontrun{
TEST_C <- c(1:3,NA,3)
is.na(TEST_C)
not.na(TEST_C)
}
}
