% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project Tools# those not functions.R
\name{not.null}
\alias{not.null}
\title{not.na}
\usage{
not.null(x)
}
\arguments{
\item{x}{quary vector}
}
\description{
not.na
}
\details{
find out items not NA
}
\examples{
\dontrun{
not.null(NULL)
not.null(3)
}
}
