% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nth_col_as_rowname#Data Process Tools#.R
\name{nth_col_as_rowname}
\alias{nth_col_as_rowname}
\title{Set nth col to rownames}
\usage{
nth_col_as_rowname(df, nth = 1, drop_before_nth = T)
}
\arguments{
\item{df}{input data.frame}

\item{nth}{default 1, the nth col will be set as new rowname.}

\item{drop_before_nth}{logical. wether drop the contains before nth col.}
}
\value{
a data.frame with new rownames
}
\description{
将数据框的第n列设定为行名。 同时可以去除/保留第n列前面的内容
}
\examples{
\dontrun{
df <- iris
set.seed(1028)
df$var3 <- rnorm(nrow(df))
df$var4 <- rnorm(nrow(df))
df \%>\% head
df \%>\% head
nth_col_as_rowname(df, 6)
nth_col_as_rowname(df, 6, drop_before_nth = F)
}
}
