% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_to_label.R
\name{p_to_label}
\alias{p_to_label}
\title{trans p to labels}
\usage{
p_to_label(
  input,
  digits = 4,
  breaks = c(-0.00000000001, 0.001, 0.01, 0.05, 1.0001),
  labels = c("***", "**", "*", ""),
  trans_to_character = T,
  drop_p = T,
  include_right_bound = F
)
}
\arguments{
\item{input}{numeric vector. p need to be transed.}

\item{digits}{the length of the digits to be reserved}

\item{breaks}{numeric vector; to cut the p value}

\item{labels}{character vector; the labels to show for different p value.}

\item{trans_to_character}{logical, indicating wether to summary those p less than 0.0001 as "<0.0001"}

\item{drop_p}{logical, indicating wether to drop p or show p in the return.}

\item{include_right_bound}{logical, default F, indicating if the intervals should be closed
on the right (and open on the left) or vice versa.}
}
\value{
return a vector of p value labels.
}
\description{
trans p to labels.
a wrap function for cut. To handle the p value transformation.
}
\examples{
\dontrun{
p_c <- c(0.001, 0.03, 0.05, 0.00001)
p_to_label(p_c)
p_to_label(p_c, drop_p = T)
}
}
