% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project Tools# get_or_set_dir.R
\name{path_create}
\alias{path_create}
\title{check and create a path.}
\usage{
path_create(path)
}
\arguments{
\item{path}{a path need to be checked.}
}
\description{
check and create a path.
}
