% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_03_Get_ROC.R
\name{plot_roc_by_pROC}
\alias{plot_roc_by_pROC}
\title{Draw the lines by using the real_I and pred_num of the ROC.}
\usage{
plot_roc_by_pROC(real_I, pred_num, fit_note = "ROC plot", IS_smooth = F)
}
\arguments{
\item{real_I}{the real group of a each sample in dataset}

\item{pred_num}{the predicted num of each sample}

\item{fit_note}{the note to be used in the filename to seperate the results}

\item{IS_smooth}{wether to smooth the ROC lines.}
}
\description{
Draw the lines by using the real_I and pred_num of the ROC.
}
\examples{
\dontrun{
pred_num <- c(0.9,0.1,0.33,0.87)
real_I <- c(0,0,1,1)
plot_roc(real_I = real_I, pred_num = pred_num)
}
}
