% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_03_Get_ROC.R
\name{plot_roc_from_module_obj}
\alias{plot_roc_from_module_obj}
\title{Get roc by module obj via pROC package.
同時繪製訓練集和驗證集的ROC曲線}
\usage{
plot_roc_from_module_obj(module_obj, IS_smooth = F, Note = "say sth")
}
\arguments{
\item{module_obj}{inheritParams from get_the_auc}

\item{Note}{control the filename in order to change the file names.}
}
\description{
Get roc by module obj via pROC package.
同時繪製訓練集和驗證集的ROC曲線
}
