% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_01_Get_module_data.R
\name{process_var_Name}
\alias{process_var_Name}
\title{process_Name for module.
清洗变量名
replace and remove sth to ensure the var name can be used correctly in a formular}
\usage{
process_var_Name(input_name, var_head = "V")
}
\arguments{
\item{input_name}{name vector :to be cleaned.}

\item{var_head}{the head for variable which wasn't started with letters.}
}
\description{
process_Name for module.
清洗变量名
replace and remove sth to ensure the var name can be used correctly in a formular
}
\examples{
\dontrun{
input_name <- c("L_Aspartic acid", "L_Serine","9987","_ddd dkkj")
process_var_Name(input_name = input_name)
}
}
