% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data Process Tools# get_text_from_screen.R
\name{Redirect_text_from_screen}
\alias{Redirect_text_from_screen}
\title{redirect some obj}
\usage{
Redirect_text_from_screen(my_message)
}
\arguments{
\item{my_message, }{obj print in the screen
and can't be assigned to other variables directly.}
}
\value{
text by line. trans imformation into text and can be value assigned to other vars.
}
\description{
(such as some model summary, which can be assigned directly.)
this function redirect the text into a temp file and reread it.
}
\examples{
\dontrun{data("iris")
head(iris)
z <- lm(Sepal.Length ~ Sepal.Width + Species, data = iris)
z
summary(z)
k <- summary(z)
k
q <- Redirect_text_from_screen(summary(z))
q
}
}
