% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data Process Tools# nth_col_as_rowname.R
\name{rownames_to_first_column}
\alias{rownames_to_first_column}
\title{Trans data.frame rowname to the first col}
\usage{
rownames_to_first_column(df, colname = "new")
}
\arguments{
\item{df}{a df}

\item{colname}{character, name of the new col for the rownames}
}
\description{
Trans data.frame rowname to the first col
}
\examples{
\dontrun{
rownames_to_first_column(iris, colname = "New col")
}
}
