% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test load data from sysdata.R
\name{Test_reread_data_from_sysdata}
\alias{Test_reread_data_from_sysdata}
\title{Test_reread_data_from_sysdata}
\usage{
Test_reread_data_from_sysdata()
}
\value{
re
}
\description{
Test_reread_data_from_sysdata
}
\details{
Test to read the data by packagename:::dataname in R/sysdata.rda.  ##
}
\examples{
\dontrun{
the_Greek_to_Enl_df <- Test_reread_data_from_sysdata()
}
}
