% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test# Test load data from sysdata.R
\name{Test_reread_data_from_sysdata2}
\alias{Test_reread_data_from_sysdata2}
\title{Test_reread_data_from_sysdata
Test to read the data by use the dataname directly in R/sysdata.rda.  ##}
\usage{
Test_reread_data_from_sysdata2()
}
\description{
Test_reread_data_from_sysdata
Test to read the data by use the dataname directly in R/sysdata.rda.  ##
}
