% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_lines#Project Tools#.R
\name{trans_lines}
\alias{trans_lines}
\title{trans mutiple lines and remove or add sth.}
\usage{
trans_lines(input_lines = readClipboard(), mode = "add #'")
}
\arguments{
\item{input_lines}{input lines. Clipboard containings as default input. ## 输入的变量}

\item{mode}{add_#', remove_#', add "," or remove ","}
}
\value{
lines with decoration.
}
\description{
将多行代码的行首 统一 添加/去除 "#' "前缀##
或者将多行代码统一 添加/去除 行末的 “,". 注意, 去除的时候，可以直接替换成空并且换行，也可以直接替换为”：”
同时，需要注意，在注释中的内容不需要换行。#
在苹果中恐怕无法正常使用。因为苹果系统的剪贴板函数略有不同。
}
