% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fast_Notes.R
\name{fast_notes}
\alias{fast_notes}
\title{Create beautiful notes conviently}
\usage{
fast_notes(word_vector)
}
\arguments{
\item{word_vector}{character vector. Words to be shown in the notes.}
}
\value{
a multi-line character. which can be used to note.
}
\description{
Create beautiful notes conviently
}
\examples{
fast_notes(c("This is a demo for note","Type your own notes here"))
}
