% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project Tools# get_or_set_dir.R
\name{get_or_set_dir}
\alias{get_or_set_dir}
\title{get or set the dir by Rstudio}
\usage{
get_or_set_dir(
  main_trunk_as_wd = F,
  source_path = "0_data_source/",
  output_path = "2_output/",
  other_dirs = NULL,
  Change_wd = T
)
}
\arguments{
\item{main_trunk_as_wd, }{Boolean, IF T, the main dir will be set as new wd.}

\item{source_path, }{character, where raw data will be stored in a project.}

\item{output_path, }{character, where the output of an analysis will be saved.}

\item{other_dirs, }{character vector with var name.Can be NULL. the name will be assigned and be returned to global env.}

\item{Change_wd, }{Boolean; Wether set source path as the working dir.}
}
\value{
return a summary about working dir and the source and output dir.
}
\description{
get or set the dir by Rstudio
}
\details{
\preformatted{Get the dir where the script were lying in.
Set the dir as main_path and create some sub file folder inside the dir.
Chose one file as the new working dir.
}
}
\examples{
\dontrun{# set the dir as wd:
get_or_set_dir(main_trunk_as_wd = T,
   source_path = "0_data_source/",
   output_path = "2_output/",
   other_dirs = c(function_path = "DIY Function/", other_dir_test = "Test dir"),
   Change_wd = F)
# set the source as wd:
get_or_set_dir(main_trunk_as_wd = F,
   source_path = "0_data_source/",
   output_path = "2_output/",
   other_dirs = c(function_path = "DIY Function/", Dir_test = "Test dir"),
   Change_wd = T)
get_or_set_dir(main_trunk_as_wd = F,
   source_path = "0_data_source/",
   output_path = "2_output/",
   other_dirs = NULL,
   Change_wd = T)
}
}
