% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_02_Build_modules_and_get_AUC.R
\name{module.Elastic_Net}
\alias{module.Elastic_Net}
\title{building modules by using LR 使用LR进行建模}
\usage{
module.Elastic_Net(module_obj, method = "repeatedcv", cv.repeats = 7)
}
\arguments{
\item{method}{resampling method}
}
\description{
building modules by using LR 使用LR进行建模
}
