% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_02_Build_modules_and_get_AUC.R
\name{module.GB}
\alias{module.GB}
\title{building models by using gbm from gbm package
\subsection{暂时有问题用不了}{

export
}}
\usage{
module.GB(module_obj)
}
\description{
building models by using gbm from gbm package
\subsection{暂时有问题用不了}{

export
}
}
