% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_02_Build_modules_and_get_AUC.R
\name{module.LR}
\alias{module.LR}
\title{building modules by using LR 使用LR进行建模}
\usage{
module.LR(module_obj)
}
\arguments{
\item{module_obj}{inheritParams from get_the_auc}
}
\description{
building modules by using LR 使用LR进行建模
}
