% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_02_Build_modules_and_get_AUC.R
\name{module.RF}
\alias{module.RF}
\title{building modules by using LR 使用LR进行建模}
\usage{
module.RF(module_obj, mtry = 3, nodesize = 5, ntree = 10000)
}
\arguments{
\item{module_obj}{inheritParams from get_the_auc}
}
\description{
building modules by using LR 使用LR进行建模
}
